subsume tohil submodules into __init__
